requirements.txt
# Core dependencies
streamlit==1.33.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
pydantic==2.5.3

# Document processing
PyMuPDF==1.23.8  # PDF processing
python-docx==1.0.1  # DOCX processing
pytesseract==0.3.10  # OCR
Pillow==10.1.0  # Image processing
nltk==3.8.1  # Text processing

# Embedding and ML
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
flair==0.13.0  # NER
vllm==0.7.3  # Optimized LLM inference
maverick-coreference==0.2.0  # Coreference resolution

# Vector and keyword search
rank-bm25==0.2.2
qdrant-client==1.7.0  # Vector database

# Utilities
tqdm==4.66.1  # Progress bars
loguru==0.7.2  # Logging
psutil==5.9.6  # System monitoring

# Relationship extraction
graphviz==0.20.1  # Visualization
networkx==3.2.1  # Graph handling
pyvis==0.3.2  # Interactive network graphs